move compat to its own file
